Reorganize cavity class to avoid circular import.
